Consider checking index out of bounds exception on initial state
I've seen implementations like this: 
```
return `You bought ${shoppingCart[0]}, ${shoppingCart[1]}, ${shoppingCart[2]}!`;
```
The initial length of the `shoppingCard` array is 2, so `shoppingCard[2]` could be out of bounds if nothing is added to the card.
Maybe this is too advanced for this stage of the curriculum. 
In case we wish to add this test, the number of the tests should be updated.